+++ Slide for Apple Watch Extension/Extensions/DateFormatter+Extensions.swift
@@ -1,0 +1,68 @@
+//
+//  DateFormatter+Extensions.swift
+//  Slide for Apple Watch Extension
+//
+//  Created by Carlos Crane on 10/3/18.
+//  Copyright © 2018 Haptic Apps. All rights reserved.
+//
+
+import Foundation
+extension DateFormatter {
+    /**
+     Formats a date as the time since that date (e.g., “Last week, yesterday, etc.”).
+     
+     - Parameter from: The date to process.
+     - Parameter numericDates: Determines if we should return a numeric variant, e.g. "1 month ago" vs. "Last month".
+     
+     - Returns: A string with formatted `date`.
+     */
+    func timeSince(from: NSDate, numericDates: Bool = false) -> String {
+        let calendar = Calendar.current
+        let now = NSDate()
+        let earliest = now.earlierDate(from as Date)
+        let latest = earliest == now as Date ? from : now
+        let components = calendar.dateComponents([.year, .day, .hour, .minute, .second], from: earliest, to: latest as Date)
+        
+        var result = ""
+        
+        if components.year! >= 2 {
+            result = "\(components.year!)y"
+        } else if components.year! >= 1 {
+            if numericDates {
+                result = "1y"
+            } else {
+                result = "Last year"
+            }
+        } else if components.day! >= 2 {
+            result = "\(components.day!)d"
+        } else if components.day! >= 1 {
+            if numericDates {
+                result = "1d"
+            } else {
+                result = "Yesterday"
+            }
+        } else if components.hour! >= 2 {
+            result = "\(components.hour!)h"
+        } else if components.hour! >= 1 {
+            if numericDates {
+                result = "1h"
+            } else {
+                result = "An hour ago"
+            }
+        } else if components.minute! >= 2 {
+            result = "\(components.minute!)m"
+        } else if components.minute! >= 1 {
+            if numericDates {
+                result = "1m"
+            } else {
+                result = "A minute ago"
+            }
+        } else if components.second! >= 3 {
+            result = "\(components.second!)s"
+        } else {
+            result = "Just now"
+        }
+        
+        return result
+    }
+}
