--- Slide for Apple Watch Extension/ScrollTable.swift
@@ -1,10 +1,0 @@
-//
-//  ScrollTable.swift
-//  Slide for Apple Watch Extension
-//
-//  Created by Carlos Crane on 9/28/18.
-//  Copyright © 2018 Haptic Apps. All rights reserved.
-//
-
-import Foundation
-import WatchKit
