+++ Slide for Apple Watch Extension/SubmissionRowController.swift
@@ -179,90 +179,3 @@ public class SubmissionRowController: NSObject {
         self.imageGroup.setCornerRadius(10)
     }
 }
-extension DateFormatter {
-    /**
-     Formats a date as the time since that date (e.g., “Last week, yesterday, etc.”).
-     
-     - Parameter from: The date to process.
-     - Parameter numericDates: Determines if we should return a numeric variant, e.g. "1 month ago" vs. "Last month".
-     
-     - Returns: A string with formatted `date`.
-     */
-    func timeSince(from: NSDate, numericDates: Bool = false) -> String {
-        let calendar = Calendar.current
-        let now = NSDate()
-        let earliest = now.earlierDate(from as Date)
-        let latest = earliest == now as Date ? from : now
-        let components = calendar.dateComponents([.year, .day, .hour, .minute, .second], from: earliest, to: latest as Date)
-        
-        var result = ""
-        
-        if components.year! >= 2 {
-            result = "\(components.year!)y"
-        } else if components.year! >= 1 {
-            if numericDates {
-                result = "1y"
-            } else {
-                result = "Last year"
-            }
-        } else if components.day! >= 2 {
-            result = "\(components.day!)d"
-        } else if components.day! >= 1 {
-            if numericDates {
-                result = "1d"
-            } else {
-                result = "Yesterday"
-            }
-        } else if components.hour! >= 2 {
-            result = "\(components.hour!)h"
-        } else if components.hour! >= 1 {
-            if numericDates {
-                result = "1h"
-            } else {
-                result = "An hour ago"
-            }
-        } else if components.minute! >= 2 {
-            result = "\(components.minute!)m"
-        } else if components.minute! >= 1 {
-            if numericDates {
-                result = "1m"
-            } else {
-                result = "A minute ago"
-            }
-        } else if components.second! >= 3 {
-            result = "\(components.second!)s"
-        } else {
-            result = "Just now"
-        }
-        
-        return result
-    }
-}
-
-extension UIImage {
-    
-    func getCopy(withSize size: CGSize) -> UIImage {
-        let hasAlpha = true
-        let scale: CGFloat = 0.0 // Use scale factor of main screen
-        let biggerSize = CGSize(width: size.width + 20, height: size.height + 20)
-        UIGraphicsBeginImageContextWithOptions(biggerSize, !hasAlpha, scale)
-        self.draw(in: CGRect(origin: CGPoint(x: 10, y: 10), size: size))
-        
-        let scaledImage = UIGraphicsGetImageFromCurrentImageContext()
-        return scaledImage!
-    }
-    
-    func getCopy(withColor color: UIColor) -> UIImage {
-        var image = withRenderingMode(.alwaysTemplate)
-        UIGraphicsBeginImageContextWithOptions(size, false, scale)
-        color.set()
-        image.draw(in: CGRect(x: 0, y: 0, width: size.width, height: size.height))
-        image = UIGraphicsGetImageFromCurrentImageContext()!
-        UIGraphicsEndImageContext()
-        return image
-    }
-    
-    func getCopy(withSize size: CGSize, withColor color: UIColor) -> UIImage {
-        return self.getCopy(withSize: size).getCopy(withColor: color)
-    }
-}
